+++ playgrounds/cloudflare/src/env.d.ts
@@ -1,4 +1,0 @@
-/// <reference path="../.astro/types.d.ts" />
-/// <reference path="../.astro/db-types.d.ts" />
-/// <reference types="astro/client" />
-/// <reference types="../.astro/virtual:astro-studio-cms.d.ts" />