+++ www/web/tailwind.config.mjs
@@ -29,5 +29,5 @@ export default {
 			},
 		},
 	},
-	plugins: [require('tailwind-scrollbar')],
+	plugins: [require('tailwind-scrollbar'), require('@tailwindcss/typography')],
 };